+++ db/migrate/20140508170532_add_randomization_requirements_to_projects.rb
@@ -1,0 +1,5 @@
+class AddRandomizationRequirementsToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :randomization_requirements, :text
+  end
+end