+++ db/migrate/20140508170635_add_attested_to_assignments.rb
@@ -1,0 +1,5 @@
+class AddAttestedToAssignments < ActiveRecord::Migration
+  def change
+    add_column :assignments, :attested, :boolean, null: false, default: false
+  end
+end