+++ test/controllers/projects_controller_test.rb
@@ -20,7 +20,7 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should create randomization for viewer" do
     login(users(:viewer))
-    post :create_randomization, id: @project, subject_code: 'S0123', stratum_1: 'BOS', stratum_2: 'Male'
+    post :create_randomization, id: @project, subject_code: 'S0123', stratum_1: 'BOS', stratum_2: 'Male', attested: '1'
     assert_not_nil assigns(:assignment)
     assert_equal 'S0123', assigns(:assignment).subject_code
     assert_equal users(:viewer), assigns(:assignment).user
@@ -33,7 +33,7 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should not create randomization for blank subject" do
     login(users(:viewer))
-    post :create_randomization, id: @project, subject_code: '', stratum_1: 'BOS', stratum_2: 'Male'
+    post :create_randomization, id: @project, subject_code: '', stratum_1: 'BOS', stratum_2: 'Male', attested: '1'
     assert_nil assigns(:assignment)
     assert assigns(:project).errors.size > 0
     assert_equal ["can't be blank"], assigns(:project).errors[:subject_code]
@@ -42,7 +42,7 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should not create randomization for existing subject" do
     login(users(:viewer))
-    post :create_randomization, id: @project, subject_code: 'S1234', stratum_1: 'BOS', stratum_2: 'Male'
+    post :create_randomization, id: @project, subject_code: 'S1234', stratum_1: 'BOS', stratum_2: 'Male', attested: '1'
     assert_nil assigns(:assignment)
     assert assigns(:project).errors.size > 0
     assert_equal ["has already been randomized"], assigns(:project).errors[:subject_code]
@@ -51,7 +51,7 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should not create randomization for list that is full" do
     login(users(:viewer))
-    post :create_randomization, id: @project, subject_code: 'S4567', stratum_1: 'BOS', stratum_2: 'Female'
+    post :create_randomization, id: @project, subject_code: 'S4567', stratum_1: 'BOS', stratum_2: 'Female', attested: '1'
     assert_nil assigns(:assignment)
     assert assigns(:project).errors.size > 0
     assert_equal ["BOS, Female is already full"], assigns(:project).errors[:list]
@@ -60,7 +60,7 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should not create randomization for list that does not exist" do
     login(users(:viewer))
-    post :create_randomization, id: @project, subject_code: 'S5566', stratum_1: 'SPR', stratum_2: 'Female'
+    post :create_randomization, id: @project, subject_code: 'S5566', stratum_1: 'SPR', stratum_2: 'Female', attested: '1'
     assert_nil assigns(:assignment)
     assert assigns(:project).errors.size > 0
     assert_equal ["SPR, Female does not exist"], assigns(:project).errors[:list]
@@ -69,10 +69,19 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should not create randomization without selected stratum" do
     login(users(:viewer))
-    post :create_randomization, id: @project, subject_code: 'S1112'
+    post :create_randomization, id: @project, subject_code: 'S1112', attested: '1'
     assert_nil assigns(:assignment)
     assert assigns(:project).errors.size > 0
     assert_equal ["must be selected"], assigns(:project).errors[:stratification_factors]
+    assert_template 'randomize_subject'
+  end
+
+  test "should not create randomization without attesting" do
+    login(users(:viewer))
+    post :create_randomization, id: @project, subject_code: 'S0123', stratum_1: 'BOS', stratum_2: 'Male', attested: '0'
+    assert_nil assigns(:assignment)
+    assert assigns(:project).errors.size > 0
+    assert_equal ["must be checked"], assigns(:project).errors[:attested]
     assert_template 'randomize_subject'
   end
 
@@ -107,10 +116,11 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should create project" do
     assert_difference('Project.count') do
-      post :create, project: { description: @project.description, name: @project.name, treatment_arms: [ { name: 'Arm One', allocation: 1 }, { name: 'Arm Two', allocation: 1 } ], block_size_multipliers: [ { value: '1', allocation: '1' }, { value: '2', allocation: '1' } ], stratification_factors: [ { name: 'Gender', options: [ 'Male', 'Female' ]}] }
+      post :create, project: { description: @project.description, name: @project.name, randomization_requirements: 'Requirements', treatment_arms: [ { name: 'Arm One', allocation: 1 }, { name: 'Arm Two', allocation: 1 } ], block_size_multipliers: [ { value: '1', allocation: '1' }, { value: '2', allocation: '1' } ], stratification_factors: [ { name: 'Gender', options: [ 'Male', 'Female' ]}] }
     end
 
     assert_equal 2, assigns(:project).treatment_arms.size
+    assert_equal 'Requirements', assigns(:project).randomization_requirements
     assert_redirected_to project_path(assigns(:project))
   end
 
@@ -146,9 +156,10 @@ class ProjectsControllerTest < ActionController::TestCase
   end
 
   test "should update project" do
-    patch :update, id: @project, project: { description: @project.description, name: @project.name, treatment_arms: [ { name: 'Arm One', allocation: 1 }, { name: 'Arm Two', allocation: 1 }, { name: 'Arm Three', allocation: 1 } ] }
+    patch :update, id: @project, project: { description: @project.description, name: @project.name, randomization_requirements: 'Requirements Updated', treatment_arms: [ { name: 'Arm One', allocation: 1 }, { name: 'Arm Two', allocation: 1 }, { name: 'Arm Three', allocation: 1 } ] }
 
     assert_equal 3, assigns(:project).treatment_arms.size
+    assert_equal 'Requirements Updated', assigns(:project).randomization_requirements
     assert_redirected_to project_path(assigns(:project))
   end